Use Promises
Switching to Promises because writing in features that take several service calls are getting too deep.  The way that the scoped http client works that hubot uses isn't as conducive to switching to Promises as I would want since you can't really leverage the Promise.denodify.

I need to circle back around and decide if I should let all errors fall through to the same catch in the promise chain and have a generic "We couldn't find what you're asking for" kind of error.